[sw] Added reset manager smoke test
